requirements.txt
python-dotenv == 1.1.0
shioaji == 1.3.0
pandas == 2.3.3
psycopg2-binary == 2.9.11
